Reorg locale package, remove helper methods from CLDR class.
Thinking from an app developer perspective: bootstrapping the framework
involves importing and constructing a global singleton instance of CLDR.
This only serves to provide instances of Engine to the application, and
it is this locale-specific instance we want to pass around.

All other functionality should be avaliable as exported types and
functions, not attached to the global singleton, so they can be imported
and used from anywhere.