Use medium dateTimeFormat to join date with time range.
When formatting an interval, if the start and end both share the same
date fields but the time differs, use a fallback format. The docs do
not specify which format to use.

In ICU this format is copied from the CLDR medium dateTimePattern as
spelled out in this comment:

 https://www.unicode.org/cldr/trac/browser/tags/release-33/tools/java/org/unicode/cldr/icu/LocaleMapper.java#L261

Related ticket: https://www.unicode.org/cldr/trac/ticket/11158