Schema customization (experimental)
This series of changes reorganizes the initialization path of the core
library such that a user can customize which identifiers are included,
and thus which values are pulled from the underlying CLDR data into the
schema.

The goal is to allow customization / exclusion of basic identifiers,
like currency codes, unit names, language / script / region codes.

This will reduce the library size slightly (including only those
identifiers that are used to build the scheual) and can significantly
reduce the size of the resource packs.