Vector arrow key fallbacks
The vector arrow now supports passing multiple keys each of which will
be tried in sequence until a non-empty value is fetched. This allows us
to encode fallbacks for alternates.

For example of day period alt ['casing', 'none'] will try 'casing' first
and if empty, fall back to 'none' which will always be defined.